Fix divergent python version for tests issue
The first line of iasimage (#!/usr/bin/env python3) indicates that the script
will use by default the python3 interpreter. That is, when the user call it
directly, after it has been installed in the PATH, e.g.

  $ iasimage create [...]

it will be interpreted using python3.

However, the tests performed here (test_iasimage.py) call iasimage explicitely
using the python2.7 interpreter [1].

This leads to confusion because on a system where python2.7 is installed and
with all the libraries up-to-date, make install will successfully perform the
tests and install iasimage, but its usage now depends on whether python3 is
installed and/or the dependencies to cryptography, etc. are met (definitely not
what the user expected after a successful install and test). For example, on
a system with no python3 installed or with missing python3 dependencies, the
above call will fail even though iasimage was validated and installed.

Since iasimage is interpreted by default using python3, the tests have been
modified to also test the script using the same version of python (so it makes
sense for the user). If the user wants to call the script with a different
version of python, but explicitely, that is a different story.

[1] https://askubuntu.com/a/475815/37468